requirements.txt
# Dependencies for agent-semantic-code microservice
# Focused on FastAPI, Ollama embeddings, and Qdrant vector storage

# --- Core Web Framework & Environment ---
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.1

# --- LangChain Ecosystem (for Ollama integration) ---
langchain==0.1.20
langchain-community==0.0.38

# --- Vector Database Clients ---
qdrant-client==1.9.0
chromadb==0.5.0
lancedb==0.6.1

# --- PyTorch for M1 GPU support ---
torch==2.3.0

# --- GitHub & Repo Processing ---
PyGithub==2.3.0
GitPython==3.1.43

# --- Document Loading & Processing ---
pypdf==4.2.0
python-docx==1.1.2
beautifulsoup4==4.12.3
Markdown==3.6

# --- Utilities ---
requests==2.32.3
tenacity==8.3.0
protobuf==4.25.3

# --- M1 Optimized Scientific Computing Libraries ---
numpy==1.26.4

# --- Testing ---
pytest==8.2.1
pytest-mock==3.14.0

# --- gRPC Tools ---
grpcio-tools==1.62.2
